Struct ego_tree::NodeId

pub struct NodeId<T> { /* fields omitted */ }

A node ID.

NodeId acts as a weak reference which is not tied to the lifetime of the Tree that created it.

With the original Tree, a NodeId can be used to obtain a NodeRef or NodeMut.

Examples

Obtaining an ID

use ego_tree::Tree;

let tree = Tree::new('a');
let root_id = tree.root().id();

Obtaining a NodeRef

let root = tree.get(root_id);