Struct ego_tree::Tree

pub struct Tree<T> { /* fields omitted */ }

A Vec-backed tree.

Nodes are allocated in a Vec which is only ever pushed to. NodeId is an opaque index into the Vec.

Each Tree has a unique ID which is also given to each NodeId it creates. This is used to bounds check a NodeId.

Methods

impl<T> Tree<T>

Important traits for Values<'a, T>
impl<'a, T: 'a> Iterator for Values<'a, T> type Item = &'a T;

pub fn values(&self) -> Values<T>

Returns an iterator over node values in creation order.

Important traits for ValuesMut<'a, T>
impl<'a, T: 'a> Iterator for ValuesMut<'a, T> type Item = &'a mut T;

pub fn values_mut(&mut self) -> ValuesMut<T>

Returns a mutable iterator over node values in creation order.

Important traits for IntoValues<T>
impl<T> Iterator for IntoValues<T> type Item = T;

pub fn into_values(self) -> IntoValues<T>

Returns an iterator that moves node values out of the tree in creation order.

Important traits for Nodes<'a, T>
impl<'a, T: 'a> Iterator for Nodes<'a, T> type Item = NodeRef<'a, T>;

pub fn nodes(&self) -> Nodes<T>

Returns an iterator over all nodes, including orphans, in creation order.

impl<T> Tree<T>

pub fn new(root: T) -> Self

Creates a new tree with a root node.

pub fn with_capacity(root: T, capacity: usize) -> Self

Creates a new tree of the specified capacity with a root node.

pub fn root(&self) -> NodeRef<T>

Returns a reference to the root node.

pub fn root_mut(&mut self) -> NodeMut<T>

Returns a mutator of the root node.

pub fn orphan(&mut self, value: T) -> NodeMut<T>

Creates an orphan node, returning a mutator of it.

pub fn get(&self, id: NodeId<T>) -> NodeRef<T>

Returns a reference to the specified node.

Panics

Panics if id does not refer to a node in this tree.

pub fn get_mut(&mut self, id: NodeId<T>) -> NodeMut<T>

Returns a mutator of the specified node.

Panics

Panics if id does not refer to a node in this tree.