Struct ego_tree::NodeMut

pub struct NodeMut<'a, T: 'a> { /* private fields */ }

Node mutator.

Implementations

impl<'a, T: 'a> NodeMut<'a, T>

pub fn id(&self) -> NodeId

Returns the ID of this node.

pub fn tree(&mut self) -> &mut Tree<T>

Returns the tree owning this node.

pub fn value(&mut self) -> &mut T

Returns the value of this node.

pub fn parent(&mut self) -> Option<NodeMut<'_, T>>

Returns the parent of this node.

pub fn prev_sibling(&mut self) -> Option<NodeMut<'_, T>>

Returns the previous sibling of this node.

pub fn next_sibling(&mut self) -> Option<NodeMut<'_, T>>

Returns the next sibling of this node.

pub fn first_child(&mut self) -> Option<NodeMut<'_, T>>

Returns the first child of this node.

pub fn last_child(&mut self) -> Option<NodeMut<'_, T>>

Returns the last child of this node.

pub fn has_siblings(&self) -> bool

Returns true if this node has siblings.

pub fn has_children(&self) -> bool

Returns true if this node has children.

pub fn append(&mut self, value: T) -> NodeMut<'_, T>

Appends a new child to this node.

pub fn prepend(&mut self, value: T) -> NodeMut<'_, T>

Prepends a new child to this node.

pub fn insert_before(&mut self, value: T) -> NodeMut<'_, T>

Inserts a new sibling before this node.

Panics

Panics if this node is an orphan.

pub fn insert_after(&mut self, value: T) -> NodeMut<'_, T>

Inserts a new sibling after this node.

Panics

Panics if this node is an orphan.

pub fn detach(&mut self)

Detaches this node from its parent.

pub fn append_id(&mut self, new_child_id: NodeId) -> NodeMut<'_, T>

Appends a child to this node.

Panics

Panics if new_child_id is not valid.

pub fn prepend_id(&mut self, new_child_id: NodeId) -> NodeMut<'_, T>

Prepends a child to this node.

Panics

Panics if new_child_id is not valid.

pub fn insert_id_before(&mut self, new_sibling_id: NodeId) -> NodeMut<'_, T>

Inserts a sibling before this node.

Panics

Panics if new_sibling_id is not valid.

Panics if this node is an orphan.

pub fn insert_id_after(&mut self, new_sibling_id: NodeId) -> NodeMut<'_, T>

Inserts a sibling after this node.

Panics

Panics if new_sibling_id is not valid.

Panics if this node is an orphan.

pub fn reparent_from_id_append(&mut self, from_id: NodeId)

Reparents the children of a node, appending them to this node.

Panics

Panics if from_id is not valid.

pub fn reparent_from_id_prepend(&mut self, from_id: NodeId)

Reparents the children of a node, prepending them to this node.

Panics

Panics if from_id is not valid.